backend/requirements.txt
# Django and DRF
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15


# Database
psycopg2-binary>=2.9.9,<3.0  # 開発環境用
# psycopg2>=2.9.9,<3.0  # 本番環境ではこちらを使用

# CORS 設定
django-cors-headers>=4.3.0,<4.4

django-extensions>=3.2

# Authentication
djangorestframework-simplejwt>=5.3.0,<6.0  # JWT 認証

# APIドキュメント
drf-spectacular>=0.26.0,<0.27

# 単体テスト
pytest>=7.0,<8.0
pytest-django>=4.0,<5.0

# requirements.txt for yomitoku OCR project
# 使用 Python 3.10+，推荐在 Anaconda 虚拟环境中运行
# 若使用 GPU 可将 onnxruntime 替换为 onnxruntime-gpu（需 CUDA 环境）
yomitoku
onnxruntime==1.17.1
numpy==1.26.4
opencv-python
Pillow